refactor(web): extract schedule utilities and optimize build chunks
- Extract formatTimeOfDay, formatNextRun, and describeSchedule from ScheduleDialog to new schedule-utils.ts module
- Remove unused ALL_MODES export from ScheduleModePicker
- Remove unused ContextMenuItem import from PluginsView
- Replace dynamic import of getSchedulerPolicy/setSchedulerPolicy with direct imports in SettingsView
- Add manual chunk splitting in vite.config for react, xterm, markdown, and lucide dependencies
- Set
